use generic types for connect and related types
prior to this there was no type-checking of connect
- the props on the resulting component are not checked to ensure you're providing the right kind of store or any other additional props
- the props on the input component are not checked to ensure the mergeProps (or map dispatch/props functions) provide all the relevant props

this means you can get errors that could be caught by the compiler but aren't

i've replaced the definitions with generic functions so that now you will have actual compile-time checks of the properties.

note that this reflects a different understanding of the wrapWithConnect function; as far as I can see, it returns a component which accepts different properties and with a state. therefore the change in the previous PR is cancelled.